Allows to remove a collection from the CollectionOptions
This required making Collection optional, as after removing it won't exist so the View will crash.  However, depending on where I unwrapped the now-optional Collection, it would either cause a crash, or an empty sheet.  Not sure why, I suspect the State change is quicker than the View diff that SwiftUI uses so things are not where they are expected to be at the point the View re-renders.  TBH sheet-on-sheet was always a questionable design so could do with a re-examination.